pyproject.toml
[project]
name = "bot-with-redis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.134.0",
    "google-genai>=1.64.0",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.41.0",
]
